woocommerce api fixed

+++ app/src/main/java/com/example/respolhpl/data/sources/ProductRepository.kt
@@ -5,5 +5,5 @@ import com.example.respolhpl.data.Result
 interface ProductRepository {
 
     suspend fun getProducts(): Result<*>
-    suspend fun getProductById(id: Long): Result<*>
+    suspend fun getProductById(id: Int): Result<*>
 }
